Enable AUTO_SELECT for more fluent typing
A character will now be committed automatically if the user continues
typing in a way that cannot be part of one transliteration. For
example, the input combination "na" could be Ꮎ, or the start of Ꮐ
("nah"). With AUTO_SELECT enabled, Ꮎ will be committed if the user
types anything but "h" next.